create term document matrix for machine learning
given a corpus or collection of documents D (represented as a list of strings) we can compute the tfidf term-document matrix which is a way of representing text as a 'bag of words' with some normalization. The function tfidf(d,D) computes the term frequency–inverse document frequency, a statistic that represents the 'importance' of a word in a corpus.